Refactor into backend and frontend
