+++ requirements.txt
@@ -3,6 +3,8 @@
 python-dotenv
 gradio
 openai>=1.0
-python-dotenv
 langdetect
+fastapi
+uvicorn
+requests
